Created Models for: Employee, Product, Store, User, Admin(inherits from User - need to add admin methods) and a joined model for Stores_Products.
Added AppDbContext file. Created DB connection and Tables.

Need to fix the following error, when seeding the DB: The value of 'Stores_Products.StoreId' is unknown when attempting to save changes. This is because the property is also part of a foreign key for which the principal entity in the relationship is not known.